fix: move server timestamp to weather-data controller
