+++ src/modules/weather-data/weather-data.controller.ts
@@ -16,6 +16,7 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+import * as moment from 'moment';
 import { ValidateGaveshaClientGuard } from '../common/guards/gavesha-client.guard';
 import { ValidateGaveshaUserGuard } from '../common/guards/gavesha-user.guard';
 import { WeatherStationsService } from '../weather-stations/weather-stations.service';
@@ -24,6 +25,7 @@ import { CreateBulkWeatherDataDto } from './dto/create-bulk-weather-data.dto';
 import { CreateWeatherComBulkWeatherDataDto } from './dto/create-weathercom-bulk-weather-data.dto';
 import { GetWeatherDatumDto } from './dto/get-weather-datum.dto';
 import { WeatherDataService } from './weather-data.service';
+
 
 @Controller('weather-data')
 @ApiTags('weather-data')
@@ -79,10 +81,11 @@ export class WeatherDataController {
       author_user_id: station.user_ids[0],
       weather_station_id: station.id,
     };
-    try {
-      const res = await this.weatherDataService.bulkCommit(dto);
+
+    let res;
+    try {    
+      res = await this.weatherDataService.bulkCommit(dto);
       console.info(res.length, 'data committed from weathercom', station.id);
-      return res;
     } catch (err) {
       console.error(
         'ERROR in bulk commit from weather computer. Request body=',
@@ -91,6 +94,21 @@ export class WeatherDataController {
       );
       throw new Error();
     }
+    const finalResponse = res.map((datum) => {
+      return {
+        _id: datum._id,
+        timestamp: datum.timestamp,
+        timestamp_iso: moment(datum.timestamp).toISOString(true),
+        created_at: (datum as any).createdAt,
+      } as BulkCreateWeatherDataResponseDto;
+    });
+
+    const returnObject = {
+       _id:null,
+       timestamp_iso: moment().toISOString(true),
+    }  as BulkCreateWeatherDataResponseDto;  
+    finalResponse.push(returnObject);
+    return finalResponse;
   }
 
   @UseGuards(ValidateGaveshaClientGuard)
